# code/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Parking foundation — local development",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/facilities/{facility_id}/commands": {
      "get": {
        "summary": "Commands",
        "operationId": "commands_api_v1_facilities__facility_id__commands_get",
        "parameters": [
          {
            "name": "facility_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Facility Id"
            }
          },
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandsHistory"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Command",
        "operationId": "command_api_v1_facilities__facility_id__commands_post",
        "parameters": [
          {
            "name": "facility_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Facility Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/facilities/{facility_id}/executions": {
      "get": {
        "summary": "Executions",
        "operationId": "executions_api_v1_facilities__facility_id__executions_get",
        "parameters": [
          {
            "name": "facility_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Facility Id"
            }
          },
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionsHistory"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{incident_id}/timeline": {
      "get": {
        "summary": "Timeline",
        "operationId": "timeline_api_v1_incidents__incident_id__timeline_get",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentTimeline"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commands/{command_id}/progress": {
      "get": {
        "summary": "Progress",
        "operationId": "progress_api_v1_commands__command_id__progress_get",
        "parameters": [
          {
            "name": "command_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Command Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandProgress"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommandInput": {
        "properties": {
          "run_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Run Id"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "query",
              "operational_goal",
              "own_vehicle_query",
              "report_exit_blocked"
            ],
            "title": "Purpose"
          },
          "text": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Text"
          },
          "target_vehicle_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Vehicle Id"
          },
          "based_on_state_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Based On State Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "purpose",
          "text",
          "based_on_state_version"
        ],
        "title": "CommandInput"
      },
...
      "CommandsHistory": {
        "properties": {
          "as_of_utc": {
            "type": "string",
            "title": "As Of Utc"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "consistency": {
            "type": "string",
            "const": "immutable_projection_current_authority",
            "title": "Consistency",
            "default": "immutable_projection_current_authority"
          },
          "preservation_limits": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Preservation Limits"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/HistoryCommand"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "as_of_utc",
          "items"
        ],
        "title": "CommandsHistory"
      },
...
      "ExecutionsHistory": {
        "properties": {
          "as_of_utc": {
            "type": "string",
            "title": "As Of Utc"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "consistency": {
            "type": "string",
            "const": "immutable_projection_current_authority",
            "title": "Consistency",
            "default": "immutable_projection_current_authority"
          },
          "preservation_limits": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Preservation Limits"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/HistoryExecution"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "as_of_utc",
          "items"
        ],
        "title": "ExecutionsHistory"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}